Optimize take for RunEndArrays
